Add wait indicator spinner
Add a spinner for deploy and purge so that the end-user gets a feedback
that it is still running. Use `github.com/caarlos0/spin` for now as this
one works with Concourse.

As a general thought, the spinner should only be used in the `internal/cmd`
package as the havener package should be independent of CLI specific details.