Improve purge behaviour
During purge of a Helm Release, collect the errors until everything is
done. In case errors are occurred, return them as a bundled error.

Improve purge of namespaces by checking whether the namespace is already
deleted (skip rest of the function), or if the namespace is already in
Phase Terminating (also, skip rest of the function).