Implement log message channel and log levels (#127)
Replaces the previous log functions with a new channel-based
system that processes the logs in the internal package. It also
introduces a log level system.

Updates the progress indicator if one is active instead of printing
the message directly.

Removes the check for interrupting signals in LogTask because
main exits the application before this code is executed.