Ensure upgrade work.
We had issues when parsing a `helm upgrade --dry-run` STDOUT. This
STDOUT need to be cleaned by two delimiters.

It seems, that for specific helm charts, the first delimiter vary
between `---\n` or `\n`. This pr fixes it, and also add in the
e2e tests, a specific chart where this delimiter variation can be observed.