Add special node exec target all
One of the common use cases for node-exec is to run a command on all
nodes of a cluster at the same time. For that, one has to first get a
list of all nodes and then specify them in the command line. Introducing
a special target called `all` delegates this task to `havener`.

Introduce special case in node lookup code to return a list of all nodes
in the cluster if the target is called `all`.

Change type of node list from pointer to the actual struct to have a
common style and less code. Remove usage of a pointer rather than the
node itself in the main node exec function code, too.

Rework the help and usage texts to reflect the special target changes.