Refine the watch output
The sorting as of the time of this commit led to some pods jumping in the list
when they became active. Also, since only some states received special coloring,
it was sometimes confusing at first glance.

Therefore, add a new routine to create a human readable output that gives the
state as either active or done. Use this in the sorting routine.

Add more cases for colors so that the overall readability improves.