Improve top command output size
The lower half of the top output did not use the full terminal height to show
as much details as possible. This was due to the fact that the box rendering
from the `neat` package always added a linefeed to the end. Therefore, the box
could not be used until the very last line.

Bump to new version of the `neat` package, which comes with a box style option
to omit the linefeed in the last line.

Rename variables that are used in the generation of the lower half output to be
more readable.

Fix calculation issue where the container per node limits used the wrong list
to setup the maximum number of lines per block.

Reduced the manual offset in the maximum line calculation to make use of the
new slimmer box outputs.