Increase number of lines in top output
The table render function always adds a linefeed at the end of the table. This
results in an empty line at the end of the lower top output command, which takes
away space that could be used to show more details.

Bump `neat` package once more to be able to use the new table style functions
that allow to omit the linefeed to further optimize the number of lines.