Add namespace filter to watch command
There are clusters with a lot of namespaces and pods, which results in a
list of pods that needs to be truncated. The resulting list might lack
the pods of interest.

Add flag to `watch` command to specify a comma separated list of
namespaces to be displayed.

Update `ListPods` function to accept optional arguments that specify the
namespaces to be looking through. If no namespace is given, `havener`
will look up all currently existing namespaces.