Improve top output container display name
If the terminal width is not very big, the lower half of the top command
cannot be displayed. This is mostly due to the fact that there is at
least one container with a very long name. Although there is a size
limit in place, this is often not enough.

Refactor code in order to provide the maximum length to the container
display function.

Calculate the maximum container display size based on the current
terminal width by limiting it to 20 % of the available size.

Bump `text` package to be able to use the `FixedLength` function that
supports an optional ellipsis. Use gray colored ellipsis in container
display function to improve readability.