Refactor top and watch command code
In order to implement new features, the existing code needs to be
cleaned up and unified as much as possible.

Introduce command settings struct and use it to store command specific
settings in `top` and `watch` command.

Sort Go imports to use the common style where the special Kubernetes
packages are placed independently.

Add comments to sections that are otherwise hard to get at first glance.