Introduce two stage node exec
The `NodeExec` command performed two tasks in one function: Prepare a pod
on the right node with the right settings and run a command in the pod that
was created. This works fine, however if multiple commands need to be
executed, a new pod has to be created for each one of them.

Introduce `preparePodOnNode` function to separate the preparation step
into a new function that can be called independently.

Clean up other function in Go source file to use pointer receiver style.