Add load average in top output
The CPU and Memory usage is not enough to decide whether a system is performing
well or not. Another useful indicator is the load average of the node.

Add usage retriever pod that runs in the background as long as `top` runs. Get
load average values using this pod as part of the details gathering.

Insert a load average display in the node usage box.

Introduce test case to check the rendering of the node usage box.