Fix top hang in case of pod scheduling issues
In case a node is cordoned, but the load on the node does not allow for a pod
placement, the top command just hangs failing to get data.

Introduce two ways that the additional data retriever Go routine exits without
gathering data so that the overall process does not hang itself.

Add special case to node details rendering so that missing data is correctly
displayed as data unavailable.