Fix container load issue with pods that do not run
A check in the container log download routine was introduced to skip
pods that are not running. However, there are occurrences where the pod
is not running, still container logs of the init containers would be
available to download.

Remove pod running check for container log download task.

Seperate container log download for init containers and containers so
that init container logs are loaded no matter the pod state.

Introduce new directory to better sort the container log files and the
downloaded filesystem files for cluster with pods that have a lot of
init containers and containers.