Introduce kind based tests
There is no common setup in `havener` to perform tests with Kubernetes
as part of the test suites. The test setup also relies on an existing
cluster to work with.

Introduce `kind` based tests so that a fresh Kubernetes cluster is
provisioned for the time of the test suite to do basic integration
tests against an actual system.

Unify the way tests are called by removing the test shell scripts to
only rely on simple `ginkgo` command calls to call both unit tests and
test cases that require a Kubernetes cluster.

Move all `helm` related tests into the end to end package.