Unify havener interface usage
There is still a mix of styles in the `havener` package with either just
functions or with pointer receiver functions that have context.

In order to improve the test cases, make all `helm` related functions
part of the `Havener` interface so that there is one common way to
initialize the Kubernetes client with a specific context.