Add wait command
When installing Helm Charts that do not support or cannot support wait
and return immediately, it is difficult to automate their deployments.

Add `wait` command that can be configured to wait and to watch a cluster
for a specific condition, for example pods starting with a common prefix
to become ready.