Improve error message of node-exec timeout
When the `node-exec` command fails to get a pod ready in time, it fails
with an error message that does not give too much details on why the pod
could not come up.

Add more details to the error message using the already existing pod
describe logic used by the `logs` command.