Fix `KUBECONFIG` default behaviour
So far, the `havener` CLI did not behave the same way as `kubectl`, when it
comes to the Kubernetes configuration usage. If the `KUBECONFIG` environment
variable is set, this should always take precedence over the command line flag
or the default system location for the Kubernetes configuration.

Unify usage of `havener.NewHavener` function call.

Introduce options to `NewHavener` function to be able to override settings.

Add check for `KUBECONFIG` environment variable.