Rework the command execution
Drop the assumtion that Stdin and TTY is the default and let the user specify
the usage of Stdin redirect and TTY setup using command-line flags.

Deprecate the `--no-tty` flag.

Update the long description of both `node-exec` and `pod-exec`.

Fix the bug that pod exec didn't work through Teleport.

Refactored the command-line flag setup and handling.

Refactored the exec functions to accept config structs.